pyproject.toml
[tool.poetry]
name = "virt-council-assistant"
version = "0.1.0"
description = "AI-powered assistant for project management via Telegram"
authors = ["Virtual Council Team"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"
pydantic = "^2.9.0"
pydantic-ai = "^0.0.14"
python-telegram-bot = "^21.6"
openai = "^1.51.0"
google-generativeai = "^0.8.0"
sqlalchemy = "^2.0.35"
alembic = "^1.13.3"
python-dotenv = "^1.0.1"
aiosqlite = "^0.20.0"
httpx = "^0.27.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"
pytest-cov = "^5.0.0"
black = "^24.10.0"
ruff = "^0.7.0"
mypy = "^1.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
